Move userHandler to handler folder, and fix
