+++ main.go
@@ -4,6 +4,7 @@ import (
 	"gorm.io/driver/mysql"
 	"gorm.io/gorm"
 	"log"
+	"silih_a3/handler"
 	"silih_a3/user"
 )
 
@@ -16,5 +17,5 @@ func main() {
 
 	userRepository := user.NewRepository(db)
 	userService := user.NewService(userRepository)
-	userHandler := user.NewUserHandler(userService)
+	userHandler := handler.NewUserHandler(userService)
 }
